move pallets files to separate directory
