+++ runtime/src/pallets/mod.rs
@@ -1,0 +1,15 @@
+pub mod appreciation;
+pub mod authorship;
+pub mod babe;
+pub mod bags_list;
+pub mod balances;
+pub mod election_provider_multi_phase;
+pub mod grandpa;
+pub mod historical;
+pub mod identity;
+pub mod session;
+pub mod staking;
+pub mod sudo;
+pub mod system;
+pub mod timestamp;
+pub mod transaction_payment;
